# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Z-Fantasy Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  /** @nullable */
  username: string | null;
  /** @nullable */
  customNickname?: string | null;
  /** @nullable */
  userTraits?: string | null;
  /** @nullable */
  activeCharacterId?: string | null;
  ticketBalance: number;
  subscriptionTier: string;
  /** @nullable */
  lastLoginTimestamp?: string | null;
  weeklyCreationsCount: number;
  dailyTriggerRequestsCount: number;
  unlockedMediaArray?: string[];
  nsfwEnabled: boolean;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  referralCode?: string | null;
}

export interface ProfileUpdate {
  customNickname?: string;
  userTraits?: string;
}

export interface NsfwToggle {
  enabled: boolean;
}

export interface TicketClaimResult {
  ticketsAdded: number;
  newBalance: number;
  nextClaimAt: string;
}

export interface ReferralInfo {
  referralLink: string;
  referralCode: string;
  referralCount?: number;
}

/**
 * @nullable
 */
export type CharacterTriggerMetadataArray = { [key: string]: unknown } | null;

export interface Character {
  characterId: string;
  /** @nullable */
  creatorId?: string | null;
  name: string;
  visibility: string;
  /** @nullable */
  systemPrompt?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  teaserDescription?: string | null;
  /** @nullable */
  initialGreeting?: string | null;
  tags: string[];
  genre: string;
  /** @nullable */
  age?: string | null;
  /** @nullable */
  triggerMetadataArray?: CharacterTriggerMetadataArray;
}

export type CharacterInputGenre = typeof CharacterInputGenre[keyof typeof CharacterInputGenre];


export const CharacterInputGenre = {
  Anime: 'Anime',
  Fantasy: 'Fantasy',
  Modern: 'Modern',
  'Sci-Fi': 'Sci-Fi',
  Dark_Goth: 'Dark Goth',
} as const;

export interface CharacterInput {
  /** @minLength 1 */
  name: string;
  age?: string;
  bio?: string;
  initialGreeting?: string;
  genre: CharacterInputGenre;
  tags?: string[];
  avatarUrl?: string;
}

export interface CharacterUpdate {
  name?: string;
  bio?: string;
  initialGreeting?: string;
  visibility?: string;
  tags?: string[];
  avatarUrl?: string;
  systemPrompt?: string;
}

export interface CharacterPage {
  items: Character[];
  total: number;
  page: number;
  limit: number;
}

export interface MessageItem {
  role: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  timestamp?: string | null;
}

export interface Conversation {
  conversationId: string;
  characterId: string;
  affectionPoints: number;
  messages: MessageItem[];
  character?: Character;
}

export interface ConversationSummary {
  conversationId: string;
  characterId: string;
  affectionPoints: number;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  lastMessageAt?: string | null;
  unread?: boolean;
  character?: Character;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface MessageReply {
  message: MessageItem;
  /** @nullable */
  imageUrl?: string | null;
  ticketsRemaining: number;
  affectionPoints?: number;
}

export interface SelfieRequest {
  description: string;
}

export interface SelfieResponse {
  imageUrl: string;
  ticketsRemaining: number;
  matched?: boolean;
}

export type GiftInputGiftType = typeof GiftInputGiftType[keyof typeof GiftInputGiftType];


export const GiftInputGiftType = {
  cyber_cocktail: 'cyber_cocktail',
  neon_bracelet: 'neon_bracelet',
  secret_key: 'secret_key',
} as const;

export interface GiftInput {
  giftType: GiftInputGiftType;
}

export interface GiftResult {
  affectionPoints: number;
  newLevel: number;
  ticketsRemaining: number;
  aiReaction: string;
  /** @nullable */
  scenarioImageUrl?: string | null;
}

export interface VaultItem {
  id: string;
  imageUrl: string;
  characterId: string;
  unlocked: boolean;
  /** @nullable */
  characterName?: string | null;
}

export interface UnlockMediaInput {
  mediaId: string;
}

export type InvoiceRequestTier = typeof InvoiceRequestTier[keyof typeof InvoiceRequestTier];


export const InvoiceRequestTier = {
  Bronze: 'Bronze',
  Silver: 'Silver',
  Gold: 'Gold',
} as const;

export type InvoiceRequestPeriod = typeof InvoiceRequestPeriod[keyof typeof InvoiceRequestPeriod];


export const InvoiceRequestPeriod = {
  weekly: 'weekly',
  monthly: 'monthly',
  yearly: 'yearly',
} as const;

export interface InvoiceRequest {
  tier: InvoiceRequestTier;
  period: InvoiceRequestPeriod;
}

export interface InvoiceResponse {
  invoiceLink: string;
}

export type PaymentWebhookPayloadPreCheckoutQuery = { [key: string]: unknown };

export type PaymentWebhookPayloadMessage = { [key: string]: unknown };

export interface PaymentWebhookPayload {
  update_id?: number;
  pre_checkout_query?: PaymentWebhookPayloadPreCheckoutQuery;
  message?: PaymentWebhookPayloadMessage;
}

export interface Transaction {
  transactionId: string;
  actionType: string;
  ticketAmount: number;
  timestamp: string;
}

export interface AdminStats {
  totalUsers: number;
  activeConversations: number;
  totalRevenue: number;
  totalCharacters: number;
  newUsersToday?: number;
  messagesLastHour?: number;
}

export interface AdminUserPage {
  items: User[];
  total: number;
  page: number;
}

export interface AdminUserDetail {
  user: User;
  transactions: Transaction[];
  conversations: ConversationSummary[];
}

export interface AdminUserEdit {
  ticketBalance?: number;
  subscriptionTier?: string;
  clearUnlockedMedia?: boolean;
  customNickname?: string;
}

export interface BroadcastInput {
  /** @minLength 1 */
  message: string;
}

export interface BroadcastResult {
  sent: number;
  failed: number;
}

export type MediaUploadInputFolder = typeof MediaUploadInputFolder[keyof typeof MediaUploadInputFolder];


export const MediaUploadInputFolder = {
  profile: 'profile',
  auto_loop: 'auto_loop',
  trigger_pool: 'trigger_pool',
  generate: 'generate',
} as const;

export interface MediaUploadInput {
  base64Data: string;
  characterId: string;
  folder: MediaUploadInputFolder;
  filename?: string;
}

export interface MediaUploadResult {
  url: string;
  publicId: string;
}

export interface SecretCheckInput {
  phrase: string;
}

export interface SecretCheckResult {
  isAdmin: boolean;
}

export type ListCharactersParams = {
page?: number;
limit?: number;
search?: string;
tags?: string;
genre?: string;
};

export type AdminListUsersParams = {
search?: string;
page?: number;
};

export type AdminListCharactersParams = {
page?: number;
};