# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Z-Fantasy Telegram Mini App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /conversations/{characterId}/messages:
    post:
      operationId: sendMessage
      tags: [conversations]
      summary: Send a message and get AI reply (costs 2 tickets for Free tier)
      parameters:
        - name: characterId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: AI reply with optional image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageReply"
        "402":
          description: Insufficient tickets or daily limit reached
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get global platform stats (admin only)
      responses:
        "200":
          description: Platform statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Forbidden
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Search/list all users (admin only)
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserPage"
        "403":
          description: Forbidden
  /admin/users/{userId}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get full user profile (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Full user data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
        "403":
          description: Forbidden

    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Admin edit user (tickets, tier, clear vault etc.)
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserEdit"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
  /admin/characters:
    get:
      operationId: adminListCharacters
      tags: [admin]
      summary: List all characters including private (admin only)
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
      responses:
        "200":
          description: All characters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterPage"
        "403":
          description: Forbidden
  /admin/characters/{characterId}/clone:
    post:
      operationId: adminCloneCharacter
      tags: [admin]
      summary: Clone a character (admin only)
      parameters:
        - name: characterId
          in: path
          required: true
          schema: { type: string }
      responses:
        "201":
          description: Cloned character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Character"
        "403":
          description: Forbidden
  /admin/broadcast:
    post:
      operationId: adminBroadcast
      tags: [admin]
      summary: Send broadcast message to all users via Telegram bot (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastInput"
      responses:
        "200":
          description: Broadcast sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BroadcastResult"
        "403":
          description: Forbidden
  /admin/upload-media:
    post:
      operationId: adminUploadMedia
      tags: [admin]
      summary: Upload media asset to Cloudinary (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaUploadInput"
      responses:
        "200":
          description: Upload result with Cloudinary URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaUploadResult"
        "403":
          description: Forbidden
  /admin/secret-check:
    post:
      operationId: adminSecretCheck
      tags: [admin]
      summary: Validate secret admin phrase and return admin panel access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecretCheckInput"
      responses:
        "200":
          description: Access granted or denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecretCheckResult"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, ticketBalance, subscriptionTier, weeklyCreationsCount, dailyTriggerRequestsCount, nsfwEnabled]
      properties:
        id: { type: string }
        username: { type: ["string", "null"] }
        customNickname: { type: ["string", "null"] }
        userTraits: { type: ["string", "null"] }
        activeCharacterId: { type: ["string", "null"] }
        ticketBalance: { type: integer }
        subscriptionTier: { type: string }
        lastLoginTimestamp: { type: ["string", "null"] }
        weeklyCreationsCount: { type: integer }
        dailyTriggerRequestsCount: { type: integer }
        unlockedMediaArray: { type: array, items: { type: string } }
        nsfwEnabled: { type: boolean }
        avatarUrl: { type: ["string", "null"] }
        referralCode: { type: ["string", "null"] }
...
    Character:
      type: object
      required: [characterId, name, visibility, tags, genre]
      properties:
        characterId: { type: string }
        creatorId: { type: ["string", "null"] }
        name: { type: string }
        visibility: { type: string }
        systemPrompt: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        teaserDescription: { type: ["string", "null"] }
        initialGreeting: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        genre: { type: string }
        age: { type: ["string", "null"] }
        triggerMetadataArray: { type: ["object", "null"] }
...
    CharacterPage:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Character"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    MessageItem:
      type: object
      required: [role, content]
      properties:
        role: { type: string }
        content: { type: string }
        imageUrl: { type: ["string", "null"] }
        timestamp: { type: ["string", "null"] }
...
    ConversationSummary:
      type: object
      required: [conversationId, characterId, affectionPoints]
      properties:
        conversationId: { type: string }
        characterId: { type: string }
        affectionPoints: { type: integer }
        lastMessage: { type: ["string", "null"] }
        lastMessageAt: { type: ["string", "null"] }
        unread: { type: boolean }
        character:
          $ref: "#/components/schemas/Character"
    MessageInput:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
    MessageReply:
      type: object
      required: [message, ticketsRemaining]
      properties:
        message:
          $ref: "#/components/schemas/MessageItem"
        imageUrl: { type: ["string", "null"] }
        ticketsRemaining: { type: integer }
        affectionPoints: { type: integer }
...
    Transaction:
      type: object
      required: [transactionId, actionType, ticketAmount, timestamp]
      properties:
        transactionId: { type: string }
        actionType: { type: string }
        ticketAmount: { type: integer }
        timestamp: { type: string }
    AdminStats:
      type: object
      required: [totalUsers, activeConversations, totalRevenue, totalCharacters]
      properties:
        totalUsers: { type: integer }
        activeConversations: { type: integer }
        totalRevenue: { type: integer }
        totalCharacters: { type: integer }
        newUsersToday: { type: integer }
        messagesLastHour: { type: integer }
    AdminUserPage:
      type: object
      required: [items, total, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total: { type: integer }
        page: { type: integer }
    AdminUserDetail:
      type: object
      required: [user, transactions, conversations]
      properties:
        user:
          $ref: "#/components/schemas/User"
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        conversations:
          type: array
          items:
            $ref: "#/components/schemas/ConversationSummary"
    AdminUserEdit:
      type: object
      properties:
        ticketBalance: { type: integer }
        subscriptionTier: { type: string }
        clearUnlockedMedia: { type: boolean }
        customNickname: { type: string }
    BroadcastInput:
      type: object
      required: [message]
      properties:
        message: { type: string, minLength: 1 }
    BroadcastResult:
      type: object
      required: [sent, failed]
      properties:
        sent: { type: integer }
        failed: { type: integer }
    MediaUploadInput:
      type: object
      required: [base64Data, characterId, folder]
      properties:
        base64Data: { type: string }
        characterId: { type: string }
        folder: { type: string, enum: [profile, auto_loop, trigger_pool, generate] }
        filename: { type: string }
    MediaUploadResult:
      type: object
      required: [url, publicId]
      properties:
        url: { type: string }
        publicId: { type: string }
    SecretCheckInput:
      type: object
      required: [phrase]
      properties:
        phrase: { type: string }
    SecretCheckResult:
      type: object
      required: [isAdmin]
      properties:
        isAdmin: { type: boolean }