# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Z-Fantasy Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  /** @nullable */
  username: string | null;
  /** @nullable */
  customNickname?: string | null;
  /** @nullable */
  userTraits?: string | null;
  /** @nullable */
  activeCharacterId?: string | null;
  ticketBalance: number;
  subscriptionTier: string;
  /** @nullable */
  lastLoginTimestamp?: string | null;
  weeklyCreationsCount: number;
  dailyTriggerRequestsCount: number;
  unlockedMediaArray?: string[];
  nsfwEnabled: boolean;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  referralCode?: string | null;
}